refactor: move useLogout to app/hooks.ts
